% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econutil-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arguments}
\alias{arguments<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{partialised}{\code{\link[partialised]{arguments}}, \code{\link[partialised:arguments]{arguments<-}}}
}}

